# app/backend/booking-service/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  internal_booking_adapter_in_http_timeoff.CreateRequest:
    properties:
      end_datetime:
        type: integer
      reason:
        type: string
      start_datetime:
        type: integer
    required:
    - end_datetime
    - reason
    - start_datetime
    type: object
info:
  contact: {}
  description: Hệ thống đặt lịch khám bệnh - MindCare
  title: Booking Service API
  version: "1.0"
paths:
...
  /booking/time-off:
    get:
      parameters:
      - description: Start date YYYY-MM-DD
        in: query
        name: from
        type: string
      - description: End date YYYY-MM-DD
        in: query
        name: to
        type: string
      - description: Processed state
        in: query
        name: processed
        type: boolean
      - description: Zero-based page
        in: query
        name: page
        type: integer
      - description: Page size, 1-100
        in: query
        name: size
        type: integer
      responses: {}
      security:
      - BearerAuth: []
      summary: '[EXPERT] Get time-off calendar'
      tags:
      - TimeOff
    post:
      consumes:
      - application/json
      description: '[EXPERT] Expert registers time-off period, returns 409 conflict
        if there are overlapping appointments'
      parameters:
      - description: Thông tin thời gian nghỉ phép
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_booking_adapter_in_http_timeoff.CreateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties: true
            type: object
        "409":
          description: Conflict
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: '[EXPERT] Register expert time-off'
      tags:
      - TimeOff
...
  /public/booking/slots/available-dates:
    get:
      parameters:
      - description: Expert ID
        in: query
        name: expert_id
        required: true
        type: string
      - description: Start date YYYY-MM-DD
        in: query
        name: from
        type: string
      - description: End date YYYY-MM-DD
        in: query
        name: to
        type: string
      - description: Zero-based page
        in: query
        name: page
        type: integer
      - description: Page size, 1-100
        in: query
        name: size
        type: integer
      responses: {}
      summary: '[PUBLIC] Get dates with bookable slots'
      tags:
      - Slots
  /public/booking/slots/available-times:
    get:
      parameters:
      - description: Expert ID
        in: query
        name: expert_id
        required: true
        type: string
      - description: Date YYYY-MM-DD
        in: query
        name: date
        required: true
        type: string
      - description: Zero-based page
        in: query
        name: page
        type: integer
      - description: Page size, 1-100
        in: query
        name: size
        type: integer
      responses: {}
      summary: '[PUBLIC] Get bookable times for a date'
      tags:
      - Slots
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"